test: restructure tests with AAA pattern and fixtures
Apply the test policy from the refactor plan:

- Move the CLI end-to-end tests out of src/main.rs into tests/cli.rs
  where cargo treats them as real integration tests; main.rs is now
  only the entry point. A tzt() fixture replaces the nine repeated
  Command::cargo_bin() calls (and migrates off the deprecated API).
- Name CLI tests after the behavior they pin down and keep the DST
  background as short Arrange comments instead of long doc blocks.
- Delete the EnvironmentVariableTzProvider test: it tested a
  crate-private struct through an env-name parameter that exists only
  for the test, and mutating process env vars races with parallel
  tests. The public provide_local_timezone_string() test still
  exercises the real lookup chain against the real system.
- Rewrite the remaining infrastructure test in AAA form.

domain/usecase/presentation tests were already written in AAA form
with fixtures in the previous PR; no changes there.
