+++ crates/infrastructure/src/get_system_timezone_from_env_var_tz.rs
@@ -20,26 +20,3 @@ impl EnvironmentVariableTzProvider {
         env::var(&self.env_name).ok()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    fn set_tmp_env_var() {
-        env::set_var("TEST_TZ", "America/New_York");
-    }
-
-    fn remove_tmp_env_var() {
-        env::remove_var("TEST_TZ");
-    }
-
-    #[test]
-    fn test_get_env_var_tz() {
-        set_tmp_env_var();
-        let tz_provider = EnvironmentVariableTzProvider::new(Some("TEST_TZ".to_string()));
-        assert_eq!(
-            tz_provider.get_env_var_tz(),
-            Some("America/New_York".to_string())
-        );
-        remove_tmp_env_var();
-    }
-}
